wheezy: gets bb-customizations, 60-omap-tty.rules is now included too

+++ tools/setup_sdcard.sh
@@ -1085,13 +1085,15 @@ populate_rootfs () {
 		cp -v ${TEMPDIR}/dl/${SPL} ${TEMPDIR}/disk/opt/backup/uboot/${spl_uboot_name}
 	fi
 
-	file="/etc/udev/rules.d/60-omap-tty.rules"
-	echo "#from: http://arago-project.org/git/meta-ti.git?a=commit;h=4ce69eff28103778508d23af766e6204c95595d3" > ${TEMPDIR}/disk${file}
-	echo "" > ${TEMPDIR}/disk${file}
-	echo "# Backward compatibility with old OMAP UART-style ttyO0 naming" > ${TEMPDIR}/disk${file}
-	echo "" >> ${TEMPDIR}/disk${file}
-	echo "SUBSYSTEM==\"tty\", ATTR{uartclk}!=\"0\", KERNEL==\"ttyS[0-9]\", SYMLINK+=\"ttyO%n\"" >> ${TEMPDIR}/disk${file}
-	echo "" >> ${TEMPDIR}/disk${file}
+	if [ ! -f ${TEMPDIR}/etc/udev/rules.d/60-omap-tty.rules ] ; then
+		file="/etc/udev/rules.d/60-omap-tty.rules"
+		echo "#from: http://arago-project.org/git/meta-ti.git?a=commit;h=4ce69eff28103778508d23af766e6204c95595d3" > ${TEMPDIR}/disk${file}
+		echo "" > ${TEMPDIR}/disk${file}
+		echo "# Backward compatibility with old OMAP UART-style ttyO0 naming" > ${TEMPDIR}/disk${file}
+		echo "" >> ${TEMPDIR}/disk${file}
+		echo "SUBSYSTEM==\"tty\", ATTR{uartclk}!=\"0\", KERNEL==\"ttyS[0-9]\", SYMLINK+=\"ttyO%n\"" >> ${TEMPDIR}/disk${file}
+		echo "" >> ${TEMPDIR}/disk${file}
+	fi
 
 	if [ "x${conf_board}" = "xam335x_boneblack" ] || [ "x${conf_board}" = "xam335x_evm" ] ; then
